refactor(poetry): command options with less duplications
